Refactor FollowButton and HomeScreen styles for improved UI consistency
- Moved FollowButton styles to a dedicated file for better organization and maintainability.
- Updated HomeScreen to utilize new styles for network error and warning icons, enhancing visual consistency.
- Removed inline styles in FollowButton for a cleaner component structure and adherence to the centralized theme.
